Divide connect code into config and connect
