+++ backend/src/mysql/connect.ts
@@ -1,11 +1,5 @@
 import { createConnection } from 'mysql'
-
-const conn = createConnection({
-  host: 'localhost',
-  user: 'root',
-  password: '1234',
-  database: 'my_db',
-})
+import conn from './config/config'
 
 conn.connect()
